# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "description" : "defaultDeription",
    "version" : "1.0.0-oas3"
  },
  "servers" : [ {
    "url" : "https://0812dce4.ngrok.io"
  } ],
  "paths" : {
...
    "/8API/List/Reports" : {
      "get" : {
        "tags" : [ "Reports" ],
        "description" : "To get the list of reports",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "hnnju2JcBL9tQd0kmLeLKKbJCX+7Hufv7+wu9TmcTV0="
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfull operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_10"
                },
                "examples" : { }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server      "
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
...
    "/8API/Reports/pagedata" : {
      "post" : {
        "tags" : [ "Reports" ],
        "description" : "To get the Report data",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "hnnju2JcBL9tQd0kmLeLKKbJCX+7Hufv7+wu9TmcTV0="
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_3"
              },
              "examples" : {
                "0" : {
                  "value" : "{\n  \"data\": [\n    {\n      \"StartingDate\": 132251905,\n      \"LanguageId\": 0,\n      \"Masters\": [\n        {\n          \"id\": 28053,\n          \"Name\": \"\"\n        }\n      ],\n      \"EndingDate\": 132318745,\n      \"LayoutId\": 9,\n      \"RowsPerPage\": 10,\n      \"CurrentPage\": 0,\n      \"ReportId\": \"551\",\n      \"Inputs\": [\n        {\n          \"id\": \"28053\",\n          \"value\": 0\n        }\n      ],\n      \"UniqueId\": \"\"\n    }\n  ]\n}\n"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "application/json; charset=utf-8"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "inline_response_200_10_data" : {
        "type" : "object",
        "properties" : {
          "ReportName" : {
            "type" : "string"
          },
          "ReportId" : {
            "type" : "integer"
          }
        },
        "example" : {
          "ReportName" : "ReportName",
          "ReportId" : 6
        }
      },
      "inline_response_200_10" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_10_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "ReportName" : "ReportName",
            "ReportId" : 6
          }, {
            "ReportName" : "ReportName",
            "ReportId" : 6
          } ],
          "message" : "message"
        }
      },
...
      "8APIReportspagedata_Masters" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "Name" : {
            "type" : "string"
          }
        }
      },
      "8APIReportspagedata_Inputs" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "value" : {
            "type" : "integer"
          }
        }
      },
      "8APIReportspagedata_data" : {
        "type" : "object",
        "properties" : {
          "StartingDate" : {
            "type" : "integer"
          },
          "Masters" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/8APIReportspagedata_Masters"
            }
          },
          "UniqueId" : {
            "type" : "string"
          },
          "EndingDate" : {
            "type" : "integer"
          },
          "CurrentPage" : {
            "type" : "integer"
          },
          "LayoutId" : {
            "type" : "integer"
          },
          "ReportId" : {
            "type" : "string"
          },
          "RowsPerPage" : {
            "type" : "integer"
          },
          "Inputs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/8APIReportspagedata_Inputs"
            }
          },
          "LanguageId" : {
            "type" : "integer"
          }
        }
      },
      "body_3" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/8APIReportspagedata_data"
            }
          }
        }
      },
...
    }
  }
}